decrease detection time
Previously the detection logic ran at a fixed cadence and waited for two
cycles before pushing a notification.
By decoupling the input/output forwarding and starting a delay as soon
as we receive the correct message we can make the time to detect a touch
wait both lower and more predictable.